pyproject.toml
[project]
name = "fishy-fishy-fishy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "environs>=14.5.0",
    "logging>=0.4.9.6",
    "python-telegram-bot==13.15",
    "redis==3.2.1",
    "requests>=2.32.5",
]
